+++ debug-queries.sql
@@ -1,0 +1,11 @@
+-- Debug Test File for Query Splitting
+-- Place cursor on different queries and execute to test
+
+-- Query 1
+select current_date;
+
+-- Query 2
+SELECT current_timestamp;
+
+-- Query 3
+SELECT current_user;